src/app/appLN/shared/pipes/format-bytes.pipe.ts
import { Pipe, PipeTransform } from '@angular/core';

@Pipe({
  name: 'formatBytes'
})
export class FormatBytesPipe implements PipeTransform {

  bytes!: number;
  decimals: number = 2;

  private units = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB'];
  transform(bytes: number = 0, precision: number = 2): string {
    if (isNaN(parseFloat(String(bytes))) || !isFinite(bytes)) return '?';
    let unit = 0;
    while (bytes >= 1024) {
      bytes /= 1024;
      unit++;
    }
    return bytes.toFixed(+precision) + this.units[unit];
  }
}
